audio: copier: move copier generic header file to copier folder
This header file is referred by external modules, add relative
path for external module usage.
