+++ src/audio/copier/copier_dai.c
@@ -6,9 +6,9 @@
 
 #include <sof/trace/trace.h>
 #include <sof/audio/component_ext.h>
-#include <ipc4/copier.h>
 #include <ipc/dai.h>
 #include <sof/audio/module_adapter/module/generic.h>
+#include "copier.h"
 #include "dai_copier.h"
 
 LOG_MODULE_DECLARE(copier, CONFIG_SOF_LOG_LEVEL);
