+++ src/ipc/ipc4/dai.c
@@ -20,13 +20,14 @@
 #include <ipc/header.h>
 #include <ipc4/alh.h>
 #include <ipc4/ssp.h>
-#include <ipc4/copier.h>
 #include <ipc4/fw_reg.h>
 #include <ipc/dai.h>
 #include <errno.h>
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
+
+#include "../audio/copier/copier.h"
 #include "../audio/copier/dai_copier.h"
 
 LOG_MODULE_DECLARE(ipc, CONFIG_SOF_LOG_LEVEL);
