+++ src/ipc/ipc4/helper.c
@@ -35,7 +35,6 @@
 #include <rimage/cavs/cavs_ext_manifest.h>
 #include <rimage/sof/user/manifest.h>
 #include <ipc4/base-config.h>
-#include <ipc4/copier.h>
 #include <ipc/header.h>
 #include <ipc4/notification.h>
 #include <ipc4/pipeline.h>
